AIModel/requirements.txt
flask
deepl
huggingface_hub
python-dotenv
requests